prepared_layers/requirements.txt
psycopg2-binary==2.9.9
pandas==2.0.3
numpy==1.24.3
schedule==1.2.1

